requirements.txt
pandas==2.1.4
numpy==1.26.2
pyspark==3.5.0
scikit-learn==1.3.2
xgboost==2.0.3
mlflow==2.9.2
pyyaml==6.0.1
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
pyarrow==14.0.1
jupyter==1.0.0
notebook==7.0.6